Align Coverage Planner chat with UI bundles
Share the Coverage Planner model between the static planner, general AI context, and the Make lab plan chat CTA. The chat prompt now carries the exact visible bundles and tells the model not to substitute generic tiers or re-add markers satisfied by equivalent core groups such as Active B12.

Checks:
- node tests/test-biology-scores.js (194 passed)
- npm run typecheck
- npm run quality
- Playwright Make lab plan prompt smoke
